// packages/src/api/generated/data-security/schemas/maskingRuleVO.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Data Security Center API (数据安全中心 API 契约)
 * 企业级数据安全与合规治理接口规范（涵盖数据分级、10级分类目录、敏感特征识别、规则调度、密钥管理、动态/静态脱敏与时效白名单）
 * OpenAPI spec version: 1.0.0
 */
import type { MaskingRuleVOAlgorithmParams } from './maskingRuleVOAlgorithmParams';
import type { MaskingRuleVOAlgorithmType } from './maskingRuleVOAlgorithmType';
import type { MaskingRuleVOApplyScene } from './maskingRuleVOApplyScene';

export interface MaskingRuleVO {
  algorithmParams?: MaskingRuleVOAlgorithmParams;
  algorithmType?: MaskingRuleVOAlgorithmType;
  applyScene?: MaskingRuleVOApplyScene;
  categoryId?: number;
  categoryName?: string;
  id?: number;
  keyId?: number;
  keyName?: string;
  ruleName?: string;
  scopeType?: string;
  status?: string;
}